refactor: remove orphaned Member Compliance UI, superseded by Observation
Personnel → Observation now covers member SSO-compliance (plus activity and
lifecycle status), so the standalone Member Compliance feature is dead code:

- delete MemberComplianceController, CorporationComplianceResource,
  routes/Routes/Corporation/MemberCompliance.php (+ its include) and the
  Corporation/MemberCompliance Vue pages
- CharacterComplianceElement is still used by the (live) corp recruitment
  ApplicationsTable, so it's relocated there rather than deleted
- ComplianceLifeCycleTest's route tests are gone; its one surviving assertion
  — that GetCorporationMemberComplianceAffiliatedIdsService resolves ids for
  CheckAuthorizationWithExtendedScope — moves to ExtendedScopeAuthorizationTest

GetCorporationMemberComplianceAffiliatedIdsService and the `view member
compliance` permission are kept: the service feeds CheckAuthorizationWithExtendedScope
(Assets/Contracts/Corp-History routes) and the permission gates Observation.
