+++ resources/js/Pages/Corporation/Recruitment/ApplicationsTable/ApplicationsTable.vue
@@ -57,7 +57,7 @@ import StickyHeaderTable from "@/Shared/Layout/Table/StickyHeaderTable.vue";
 import StickyHeaderTableRow from "@/Shared/Layout/Table/StickyHeaderTableRow.vue";
 import StickyHeaderCell from "@/Shared/Layout/Table/StickyHeaderCell.vue";
 import EntityByIdBlock from "@/Shared/Layout/Eve/EntityByIdBlock.vue";
-import CharacterComplianceElement from "@/Pages/Corporation/MemberCompliance/CharacterComplianceElement.vue";
+import CharacterComplianceElement from "./CharacterComplianceElement.vue";
 import Button from "@/Shared/Layout/Button.vue";
 
 let headerTitles = [
