+++ resources/js/Pages/Corporation/Recruitment/ApplicationsTable/CharacterComplianceElement.vue
