--- routes/Routes/Corporation/MemberCompliance.php
@@ -1,51 +1,0 @@
-<?php
-
-/*
- * MIT License
- *
- * Copyright (c) 2019, 2020, 2021 Felix Huber
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-use Illuminate\Support\Facades\Route;
-use Seatplus\Auth\Http\Middleware\CheckAuthorization;
-use Seatplus\Web\Http\Controllers\Corporation\MemberCompliance\MemberComplianceController;
-
-Route::prefix('compliance')
-    ->group(function () {
-
-        Route::middleware(CheckAuthorization::class.':view member compliance,director')
-            ->group(function () {
-
-                Route::get('', [MemberComplianceController::class, 'index'])->name('corporation.member_compliance');
-            });
-
-        Route::middleware(CheckAuthorization::class.':view member compliance,director')
-            ->group(function () {
-
-                Route::get('/{corporation_id}/{type}', [MemberComplianceController::class, 'getCorporationCompliance'])
-                    ->name('corporation.compliance');
-
-                Route::get('{corporation_id}/review/{user}', [MemberComplianceController::class, 'reviewUser'])
-                    ->middleware(CheckAuthorization::class.':member compliance: review user')
-                    ->name('corporation.review.user');
-            });
-
-    });
