+++ routes/routes.php
@@ -85,7 +85,6 @@ Route::middleware('web')
                         include __DIR__.'/Routes/Corporation/Wallet.php';
                         include __DIR__.'/Routes/Corporation/MemberTracking.php';
                         include __DIR__.'/Routes/Corporation/Recruitment.php';
-                        include __DIR__.'/Routes/Corporation/MemberCompliance.php';
                     });
 
                 Route::prefix('recruitment')
