--- tests/Feature/ComplianceLifeCycleTest.php
@@ -1,337 +1,0 @@
-<?php
-
-use Illuminate\Support\Facades\Event;
-use Inertia\Testing\AssertableInertia as Assert;
-use Seatplus\Auth\Models\CharacterUser;
-use Seatplus\Auth\Models\Permissions\Permission;
-use Seatplus\Auth\Models\User;
-use Seatplus\Eveapi\Models\Character\CharacterRole;
-use Seatplus\Eveapi\Models\Corporation\CorporationInfo;
-use Seatplus\Eveapi\Models\SsoScopes;
-use Seatplus\Web\Services\Affiliations\GetCorporationMemberComplianceAffiliatedIdsService;
-
-beforeEach(function () {
-    test()->secondary_user = Event::fakeFor(fn () => User::factory()->create());
-
-    test()->superuser = Event::fakeFor(function () {
-        $user = User::factory()->create();
-
-        $permission = Permission::findOrCreate('superuser');
-
-        $user->givePermissionTo($permission);
-
-        // now re-register all the roles and permissions
-
-        return $user;
-    });
-
-    test()->secondary_character = test()->secondary_user->characters->first();
-
-    // Ensure secondary_user's character has no EVE corporation roles (e.g. Director).
-    // CharacterInfoFactory creates CharacterRole with roles=[] by default, but stale records
-    // or Event::fakeFor interference can leave a role with Director, causing permission bypasses.
-    CharacterRole::updateOrCreate(
-        ['character_id' => test()->secondary_character->character_id],
-        ['roles' => [], 'roles_at_base' => null, 'roles_at_hq' => null, 'roles_at_other' => null]
-    );
-});
-
-test('user without permission fails to see compliance', function () {
-    if (test()->test_user->can('superuser')) {
-        test()->test_user->removeRole('superuser');
-
-        // now re-register all the roles and permissions
-    }
-
-    $response = test()->actingAs(test()->secondary_user)
-        ->get(route('corporation.member_compliance'))
-        ->assertForbidden();
-});
-
-test('user with permission sees component', function () {
-    if (test()->test_user->can('superuser')) {
-        test()->test_user->removeRole('superuser');
-
-        // now re-register all the roles and permissions
-    }
-
-    $response = test()->actingAs(test()->test_user)
-        ->get(route('corporation.member_compliance'))
-        ->assertForbidden();
-
-    assignPermissionToTestUser(['view member compliance']);
-
-    $response = test()->actingAs(test()->test_user)
-        ->get(route('corporation.member_compliance'))
-        ->assertOk();
-
-    $response->assertInertia(fn (Assert $page) => $page->component('Corporation/MemberCompliance/MemberCompliance'));
-});
-
-test('user with permission sees default compliance', function () {
-    createScopeSetting();
-
-    test()->withoutMiddleware();
-
-    $response = test()->actingAs(test()->test_user)
-        ->getJson(route('corporation.compliance', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'type' => 'default',
-        ]))
-        ->assertOk();
-
-    $response->assertJsonCount(1, 'data');
-});
-
-it('is possible to search for a character', function () {
-    createScopeSetting();
-
-    test()->withoutMiddleware();
-
-    $response = test()->actingAs(test()->test_user)
-        ->getJson(route('corporation.compliance', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'type' => 'default',
-            'search' => substr((string) test()->secondary_character->name, 5),
-        ]))
-        ->assertOk();
-
-    $response->assertJsonCount(1, 'data');
-});
-
-test('user with permission sees user compliance', function () {
-    createScopeSetting(['view member compliance'], 'user');
-
-    test()->withoutMiddleware();
-
-    $response = test()->actingAs(test()->test_user)
-        ->getJson(route('corporation.compliance', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'type' => 'user',
-        ]));
-
-    $response->assertJsonCount(1, 'data');
-
-    expect(test()->test_user->characters)->toHaveCount(1);
-
-    $response->assertJsonFragment(['count_total' => 1]);
-
-    CharacterUser::query()->where('character_id', test()->secondary_character->character_id)
-        ->update(['user_id' => test()->test_user->id]);
-
-    expect(test()->test_user->refresh()->characters)->toHaveCount(2);
-
-    $response = test()->actingAs(test()->test_user)
-        ->getJson(route('corporation.compliance', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'type' => 'user',
-        ]));
-
-    $response->assertJsonFragment(['count_total' => 2]);
-});
-
-test('non director can not access the compliance index', function () {
-    // 1. non director can't access the compliance index
-    $non_director = Event::fakeFor(function () {
-        $user = User::factory()->create();
-
-        $roles = $user->characters->first()->roles;
-        $roles->roles = ['Contract_Manager'];
-        $roles->save();
-
-        return $user->refresh();
-    });
-
-    test()->actingAs($non_director)
-        ->get(route('corporation.member_compliance'))
-        ->assertForbidden();
-});
-
-test('director user without permission can access index', function () {
-    $director = Event::fakeFor(function () {
-        $user = User::factory()->create();
-
-        $roles = $user->characters->first()->roles;
-        $roles->roles = ['Director'];
-        $roles->save();
-
-        return $user->refresh();
-    });
-
-    $response = test()->actingAs($director)
-        ->get(route('corporation.member_compliance'))
-        ->assertOk();
-
-    $response->assertInertia(
-        fn (Assert $page) => $page
-            ->component('Corporation/MemberCompliance/MemberCompliance')
-    );
-});
-
-test('director user without permission can review its corp members', function () {
-    // createScopeSetting();
-
-    // 1. make sure user ist not superuser
-    expect(test()->test_user->can('superuser'))->toBeFalse();
-
-    // 2. director can access the compliance index
-    $director = Event::fakeFor(function () {
-        $user = test()->test_user;
-
-        $roles = $user->characters->first()->roles;
-        $roles->roles = ['Director'];
-        $roles->save();
-
-        return $user->refresh();
-    });
-
-    expect(test()->test_character->load('roles')->roles->roles)->toContain('Director');
-
-    // 3. setup ssoScope
-    SsoScopes::updateOrCreate([
-        'morphable_id' => test()->test_character->corporation->corporation_id,
-    ], [
-        'selected_scopes' => test()->test_character->refreshToken->scopes,
-        'morphable_type' => CorporationInfo::class,
-        'type' => 'default',
-    ]);
-
-    CorporationInfo::query()
-        ->has('ssoScopes')
-        ->orHas('alliance.ssoScopes')
-        ->select('name', 'corporation_id')
-        ->addSelect(['type' => SsoScopes::whereColumn('morphable_id', 'corporation_id')->limit(1)->select('type')])
-        ->get();
-
-    $response = test()->actingAs($director)
-        ->get(route('corporation.member_compliance'))
-        ->assertOk();
-
-    $response->assertInertia(
-        fn (Assert $page) => $page
-            ->component('Corporation/MemberCompliance/MemberCompliance')
-            ->has('corporations', 1)
-    );
-});
-
-it('enables superuser to review corporation member', function () {
-    createScopeSetting();
-
-    expect(test()->superuser)
-        ->can('superuser')->toBeTrue();
-
-    $response = test()->actingAs(test()->superuser)
-        ->getJson(route('corporation.review.user', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'user' => test()->test_user->id,
-        ]))
-        ->assertOk()
-        ->assertInertia(fn (Assert $page) => $page->component('Corporation/MemberCompliance/ReviewUser'));
-});
-
-it('enables with review permission to review corporation member', function () {
-    createScopeSetting(['view member compliance', 'member compliance: review user']);
-
-    expect(test()->test_user)
-        ->can('superuser')->toBeFalse()
-        ->can('member compliance: review user')->toBeTrue()
-        ->can('view member compliance')->toBeTrue();
-
-    $response = test()->actingAs(test()->test_user)
-        ->getJson(route('corporation.review.user', [
-            'corporation_id' => test()->secondary_character->corporation->corporation_id,
-            'user' => test()->test_user->id,
-        ]))->assertOk()
-        ->assertInertia(fn (Assert $page) => $page->component('Corporation/MemberCompliance/ReviewUser'));
-});
-
-it('allows user with review permission to review corporation member', function () {
-    Event::fake();
-
-    // create a user with two characters
-    $user = User::factory()->create();
-    $secondary_character = CharacterUser::factory()->make();
-
-    $user->characterUsers()->save($secondary_character);
-
-    expect($user->characters->count())->toEqual(2);
-
-    $first_character = $user->characters()->with('corporation')->first();
-    $second_character = $user->characters()->with('corporation')->get()->last();
-
-    expect($first_character->character_id)->not()->toEqual($second_character->character_id);
-
-    // create role with affiliation and permission via HTTP
-    createRoleViaHttp(
-        roleName: faker()->name(),
-        affiliations: [
-            [
-                'entity_id' => $first_character->character_id,
-                'entity_type' => 'character',
-                'affiliation_type' => 'allowed',
-            ],
-        ],
-        member: test()->test_user,
-        permissions: ['member compliance: review user'],
-    );
-
-    // check if test user has permission
-    expect(test()->test_user->refresh()->can('member compliance: review user'))->toBeTrue();
-
-    // create sso scope
-    $sso_scope = SsoScopes::factory()->create([
-        'morphable_type' => CorporationInfo::class,
-        'morphable_id' => $first_character->corporation->corporation_id,
-        'type' => 'user',
-        'selected_scopes' => collect(['esi-alliances.read_corporations.v1'])->toJson(),
-    ]);
-
-    \Pest\Laravel\actingAs(test()->test_user);
-    $affiliated_ids = GetCorporationMemberComplianceAffiliatedIdsService::make()->getQuery()->get();
-    $affiliated_character_ids = $affiliated_ids->pluck('affiliated_id');
-
-    expect($affiliated_ids)->toHaveCount(2)
-        ->and($affiliated_character_ids)->toContain($first_character->character_id)
-        ->and($affiliated_character_ids)->toContain($second_character->character_id);
-
-    // The reviewer can reach the character skills page. Per-character skills now load as
-    // affiliation-scoped Inertia deferred props on the index route rather than a dedicated
-    // JSON endpoint, so we assert the page renders rather than hitting a removed route.
-    $response = test()->actingAs(test()->test_user)->get(route('character.skills'));
-
-    $response->assertOk();
-});
-
-// Helpers
-function createScopeSetting(array $permissons = [], $type = 'default')
-{
-    createRoleViaHttp(
-        roleName: 'test',
-        affiliations: [
-            [
-                'entity_id' => test()->secondary_character->corporation->corporation_id,
-                'entity_type' => 'corporation',
-                'affiliation_type' => 'allowed',
-            ],
-        ],
-        member: test()->test_user,
-        permissions: $permissons,
-    );
-
-    expect(test()->test_user->refresh()->hasRole('test'))->toBeTrue();
-
-    expect(SsoScopes::all())->toBeEmpty();
-
-    // Make sure secondary character is missing the required scope
-    expect(in_array('esi-assets.read_assets.v1', test()->secondary_character->refreshToken->scopes))->toBeFalse();
-
-    SsoScopes::updateOrCreate([
-        'morphable_id' => test()->secondary_character->corporation->corporation_id,
-    ], [
-        'selected_scopes' => ['esi-assets.read_assets.v1', 'esi-universe.read_structures.v1'],
-        'morphable_type' => CorporationInfo::class,
-        'type' => $type,
-    ]);
-
-    expect(SsoScopes::all())->toHaveCount(1);
-}
